requirements.txt
pandas>=2.0.0
numpy>=1.24.0
PyPortfolioOpt>=1.5.0
requests>=2.28.0
matplotlib>=3.6.0
seaborn>=0.11.0
python-dotenv>=0.19.0
yfinance>=0.2.0
scipy>=1.9.0
plotly>=5.0.0
